test: extract runtime engine suites
