+++ crates/agent-engine/src/runtime/engine.rs
@@ -1374,1443 +1374,5 @@ fn spawn_with_prepared_runtime_environment(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::runtime::{agent_loop::DeferredRuntimeAction, memory_promotion};
-    use alan_agent_protocol::{ContentPart, Op};
-    use alan_ap::InProcessTransport;
-    use alan_llm::{
-        GenerationRequest, GenerationResponse, LlmProvider, MockLlmProvider, StreamChunk,
-        TokenUsage, ToolCallDelta,
-    };
-    use anyhow::anyhow;
-    use async_trait::async_trait;
-    use std::path::Path;
-    use std::sync::{Arc, Mutex};
-    use tempfile::TempDir;
-
-    struct PackageTestTool {
-        name: &'static str,
-        description: &'static str,
-    }
-
-    impl crate::tools::Tool for PackageTestTool {
-        fn name(&self) -> &str {
-            self.name
-        }
-
-        fn description(&self) -> &str {
-            self.description
-        }
-
-        fn parameters_schema(&self) -> serde_json::Value {
-            serde_json::json!({"type": "object"})
-        }
-
-        fn execute(
-            &self,
-            _arguments: serde_json::Value,
-            _ctx: &crate::tools::ToolContext,
-        ) -> crate::tools::ToolResult {
-            Box::pin(async { Ok(serde_json::json!({"ok": true})) })
-        }
-    }
-
-    fn single_file_fs(name: &str, bytes: &[u8]) -> Arc<alan_ap::reference::MemFs> {
-        Arc::new(alan_ap::reference::MemFs::with_read_only_file(name, bytes))
-    }
-
-    fn namespace_environment_for_test() -> NamespaceRuntimeEnvironment {
-        let root = InProcessTransport::new(Arc::new(alan_kernel::MountFs::new(
-            alan_kernel::Namespace::new(),
-        )));
-        crate::runtime::NamespaceRuntimeEnvironment::new(root, "/agent/1", "default")
-    }
-
-    fn write_agent_overlay(path: &Path, body: &str) {
-        std::fs::write(path, body).unwrap();
-    }
-
-    fn make_deferred_action_for_test() -> DeferredRuntimeAction {
-        let temp = TempDir::new().unwrap();
-        let memory_dir = temp.path().join("memory-store");
-
-        let mut machine = AgentMachine::new();
-        machine.add_user_message("My name is Morris.");
-
-        let mut turn_state = TurnState::default();
-        turn_state.begin_turn(0);
-
-        let mut core_config = crate::Config::default();
-        core_config.memory.enabled = true;
-        core_config.memory.store_dir = Some(memory_dir);
-        let runtime_config = RuntimeConfig::from(&core_config);
-
-        let state = RuntimeLoopState {
-            machine,
-            current_submission_id: None,
-            environment: namespace_environment_for_test(),
-            core_config,
-            runtime_config,
-            definition_persona_dirs: Vec::new(),
-            prompt_cache: crate::runtime::prompt_cache::PromptAssemblyCache::new(Vec::new()),
-            turn_state,
-        };
-
-        memory_promotion::build_turn_memory_promotion_job(&state, "queue ordering test")
-            .map(DeferredRuntimeAction::TurnMemoryPromotion)
-            .expect("build deferred memory promotion job")
-    }
-
-    fn queue_item_kinds(queue: &VecDeque<QueuedRuntimeItem>) -> Vec<&'static str> {
-        queue
-            .iter()
-            .map(|item| match item {
-                QueuedRuntimeItem::Submission(_) => "submission",
-                QueuedRuntimeItem::Deferred(_) => "deferred",
-            })
-            .collect()
-    }
-
-    #[tokio::test]
-    async fn namespace_discovery_ignores_incomplete_tool_packages() {
-        let manifest = crate::runtime::ToolPackageManifest::from_tool(
-            &PackageTestTool {
-                name: "hidden",
-                description: "Hidden Tool",
-            },
-            30,
-        )
-        .unwrap();
-        let manifest_fs = single_file_fs("manifest", &serde_json::to_vec(&manifest).unwrap());
-
-        let mut mounts = alan_kernel::Namespace::new();
-        mounts.mount(
-            "/bin/ordinary",
-            InProcessTransport::new(Arc::new(alan_ap::reference::MemFs::new())),
-            alan_kernel::Access::ReadOnly,
-        );
-        mounts.mount(
-            "/lib/exec/hidden",
-            InProcessTransport::new(manifest_fs),
-            alan_kernel::Access::ReadOnly,
-        );
-        let root = InProcessTransport::new(Arc::new(alan_kernel::MountFs::new(mounts)));
-        let environment =
-            crate::runtime::NamespaceRuntimeEnvironment::new(root, "/agent/1", "default");
-
-        assert!(
-            environment
-                .discover_tool_packages()
-                .await
-                .unwrap()
-                .is_empty()
-        );
-    }
-
-    #[tokio::test]
-    async fn namespace_discovery_rejects_invalid_mounted_manifest() {
-        let mut mounts = alan_kernel::Namespace::new();
-        mounts.mount(
-            "/bin/broken",
-            InProcessTransport::new(Arc::new(alan_ap::reference::MemFs::new())),
-            alan_kernel::Access::ReadOnly,
-        );
-        mounts.mount(
-            "/lib/exec/broken",
-            InProcessTransport::new(single_file_fs("manifest", b"{}")),
-            alan_kernel::Access::ReadOnly,
-        );
-        let root = InProcessTransport::new(Arc::new(alan_kernel::MountFs::new(mounts)));
-        let environment =
-            crate::runtime::NamespaceRuntimeEnvironment::new(root, "/agent/1", "default");
-
-        assert!(environment.discover_tool_packages().await.is_err());
-    }
-
-    #[test]
-    fn test_should_requeue_deferred_action_only_after_cancelled_exit() {
-        assert!(should_requeue_deferred_action(
-            true,
-            DeferredRuntimeActionExit::Cancelled
-        ));
-        assert!(!should_requeue_deferred_action(
-            true,
-            DeferredRuntimeActionExit::Completed
-        ));
-        assert!(!should_requeue_deferred_action(
-            false,
-            DeferredRuntimeActionExit::Cancelled
-        ));
-    }
-
-    #[test]
-    fn paused_submission_keeps_its_terminal_ui_state() {
-        assert!(preserves_paused_terminal_state(&Ok(()), true));
-        assert!(!preserves_paused_terminal_state(&Ok(()), false));
-        assert!(!preserves_paused_terminal_state(
-            &Err(anyhow!("failed")),
-            true
-        ));
-    }
-
-    fn mock_generation_response(content: impl Into<String>) -> GenerationResponse {
-        GenerationResponse {
-            content: content.into(),
-            thinking: None,
-            thinking_signature: None,
-            redacted_thinking: Vec::new(),
-            tool_calls: Vec::new(),
-            usage: Some(TokenUsage {
-                prompt_tokens: 10,
-                cached_prompt_tokens: None,
-                completion_tokens: 5,
-                total_tokens: 15,
-                reasoning_tokens: None,
-            }),
-            finish_reason: None,
-            provider_response_id: None,
-            provider_response_status: None,
-            warnings: Vec::new(),
-        }
-    }
-
-    async fn wait_for_ui_turn_completion(
-        tail: &mut alan_shell::Tail,
-        timeout: Duration,
-    ) -> Vec<alan_agent_protocol::UiEvent> {
-        tokio::time::timeout(timeout, async {
-            let mut pending = String::new();
-            let mut events = Vec::new();
-            let mut saw_running = false;
-            loop {
-                pending.push_str(&String::from_utf8(tail.read(4096).await.unwrap()).unwrap());
-                while let Some(newline) = pending.find('\n') {
-                    let line = pending[..newline].to_string();
-                    pending.drain(..=newline);
-                    let event: alan_agent_protocol::UiEvent = serde_json::from_str(&line).unwrap();
-                    if let alan_agent_protocol::UiEvent::Activity { snapshot } = &event {
-                        saw_running |= matches!(
-                            snapshot.state,
-                            alan_agent_protocol::UiActivityState::Running
-                        );
-                        if saw_running
-                            && matches!(snapshot.state, alan_agent_protocol::UiActivityState::Idle)
-                        {
-                            events.push(event);
-                            return events;
-                        }
-                    }
-                    events.push(event);
-                }
-            }
-        })
-        .await
-        .expect("turn UI stream did not reach idle")
-    }
-
-    fn response_stream(response: GenerationResponse) -> tokio::sync::mpsc::Receiver<StreamChunk> {
-        let (tx, rx) = tokio::sync::mpsc::channel(16);
-        tokio::spawn(async move {
-            if !response.content.is_empty()
-                || response
-                    .thinking
-                    .as_deref()
-                    .is_some_and(|value| !value.is_empty())
-                || response
-                    .thinking_signature
-                    .as_deref()
-                    .is_some_and(|value| !value.is_empty())
-                || !response.redacted_thinking.is_empty()
-            {
-                let mut redacted = response.redacted_thinking.into_iter();
-                let _ = tx
-                    .send(StreamChunk {
-                        text: (!response.content.is_empty()).then_some(response.content),
-                        thinking: response.thinking,
-                        thinking_signature: response.thinking_signature,
-                        redacted_thinking: redacted.next(),
-                        usage: None,
-                        provider_response_id: None,
-                        provider_response_status: None,
-                        sequence_number: None,
-                        tool_call_delta: None,
-                        is_finished: false,
-                        finish_reason: None,
-                    })
-                    .await;
-                for redacted in redacted {
-                    let _ = tx
-                        .send(StreamChunk {
-                            text: None,
-                            thinking: None,
-                            thinking_signature: None,
-                            redacted_thinking: Some(redacted),
-                            usage: None,
-                            provider_response_id: None,
-                            provider_response_status: None,
-                            sequence_number: None,
-                            tool_call_delta: None,
-                            is_finished: false,
-                            finish_reason: None,
-                        })
-                        .await;
-                }
-            }
-
-            let tool_calls = response.tool_calls;
-            for (index, tool_call) in tool_calls.iter().enumerate() {
-                let arguments =
-                    serde_json::to_string(&tool_call.arguments).unwrap_or_else(|_| "{}".into());
-                let _ = tx
-                    .send(StreamChunk {
-                        text: None,
-                        thinking: None,
-                        thinking_signature: None,
-                        redacted_thinking: None,
-                        usage: None,
-                        provider_response_id: None,
-                        provider_response_status: None,
-                        sequence_number: None,
-                        tool_call_delta: Some(ToolCallDelta {
-                            index,
-                            id: tool_call.id.clone(),
-                            name: Some(tool_call.name.clone()),
-                            arguments_delta: Some(arguments.clone()),
-                            arguments: Some(arguments),
-                        }),
-                        is_finished: false,
-                        finish_reason: None,
-                    })
-                    .await;
-            }
-
-            let finish_reason = response.finish_reason.unwrap_or_else(|| {
-                if tool_calls.is_empty() {
-                    "stop".to_string()
-                } else {
-                    "tool_calls".to_string()
-                }
-            });
-            let _ = tx
-                .send(StreamChunk {
-                    text: None,
-                    thinking: None,
-                    thinking_signature: None,
-                    redacted_thinking: None,
-                    usage: response.usage,
-                    provider_response_id: response.provider_response_id,
-                    provider_response_status: response.provider_response_status,
-                    sequence_number: None,
-                    tool_call_delta: None,
-                    is_finished: true,
-                    finish_reason: Some(finish_reason),
-                })
-                .await;
-        });
-        rx
-    }
-
-    struct ShutdownDrainMemoryPromotionProvider {
-        call_count: Arc<Mutex<usize>>,
-        deferred_delay: Duration,
-    }
-
-    #[async_trait]
-    impl LlmProvider for ShutdownDrainMemoryPromotionProvider {
-        async fn generate(
-            &mut self,
-            _request: GenerationRequest,
-        ) -> anyhow::Result<GenerationResponse> {
-            let current_call = {
-                let mut guard = self.call_count.lock().unwrap();
-                let current = *guard;
-                *guard += 1;
-                current
-            };
-
-            match current_call {
-                0 => Ok(mock_generation_response("Noted.")),
-                1 => {
-                    tokio::time::sleep(self.deferred_delay).await;
-                    Ok(mock_generation_response(
-                        serde_json::json!({
-                            "writes": [
-                                {
-                                    "kind": "user_identity",
-                                    "target": "USER.md",
-                                    "confidence": "high",
-                                    "disposition": "promote_now",
-                                    "observation": "Name: Morris",
-                                    "evidence": ["My name is Morris."],
-                                    "promotion_rationale": "Direct user-stated stable identity detail."
-                                }
-                            ]
-                        })
-                        .to_string(),
-                    ))
-                }
-                _ => Ok(mock_generation_response(
-                    serde_json::json!({ "writes": [] }).to_string(),
-                )),
-            }
-        }
-
-        async fn chat(&mut self, _system: Option<&str>, _user: &str) -> anyhow::Result<String> {
-            Err(anyhow!(
-                "ShutdownDrainMemoryPromotionProvider does not implement chat"
-            ))
-        }
-
-        async fn generate_stream(
-            &mut self,
-            request: GenerationRequest,
-        ) -> anyhow::Result<tokio::sync::mpsc::Receiver<StreamChunk>> {
-            Ok(response_stream(self.generate(request).await?))
-        }
-
-        fn provider_name(&self) -> &'static str {
-            "shutdown_drain_memory_promotion"
-        }
-    }
-
-    #[test]
-    fn test_push_outer_submission_inserts_before_existing_deferred_actions() {
-        let mut queues = RuntimeSubmissionQueues::default();
-
-        let first_submission = Submission::new(Op::Interrupt);
-        let second_submission = Submission::new(Op::CompactWithOptions { focus: None });
-        let first_submission_id = first_submission.id.clone();
-        let second_submission_id = second_submission.id.clone();
-
-        queues.push_outer_submission(first_submission);
-        queues.push_outer_deferred(make_deferred_action_for_test());
-        queues.push_outer_deferred(make_deferred_action_for_test());
-        queues.push_outer_submission(second_submission);
-
-        assert_eq!(
-            queue_item_kinds(&queues.outer_queue),
-            vec!["submission", "submission", "deferred", "deferred"]
-        );
-
-        let queued_submission_ids = queues
-            .outer_queue
-            .iter()
-            .filter_map(|item| match item {
-                QueuedRuntimeItem::Submission(submission) => Some(submission.id.clone()),
-                QueuedRuntimeItem::Deferred(_) => None,
-            })
-            .collect::<Vec<_>>();
-        assert_eq!(
-            queued_submission_ids,
-            vec![first_submission_id, second_submission_id]
-        );
-    }
-
-    #[tokio::test]
-    async fn test_requeue_active_turn_leftovers_inserts_before_existing_deferred_actions() {
-        let mut queues = RuntimeSubmissionQueues::default();
-        queues.push_outer_deferred(make_deferred_action_for_test());
-
-        let mut turn_state = TurnState::default();
-        let buffered_submission = Submission::new(Op::Input {
-            parts: vec![alan_agent_protocol::ContentPart::text("follow up")],
-            mode: alan_agent_protocol::InputMode::FollowUp,
-        });
-        let buffered_submission_id = buffered_submission.id.clone();
-        turn_state.push_buffered_inband_submission(buffered_submission);
-
-        let requeued = queues.requeue_active_turn_leftovers(&mut turn_state).await;
-
-        assert_eq!(requeued, 1);
-        assert_eq!(
-            queue_item_kinds(&queues.outer_queue),
-            vec!["submission", "deferred"]
-        );
-
-        match queues.outer_queue.front() {
-            Some(QueuedRuntimeItem::Submission(submission)) => {
-                assert_eq!(submission.id, buffered_submission_id);
-            }
-            _ => panic!("expected buffered submission at queue front"),
-        }
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn test_runtime_shutdown_drains_deferred_memory_promotion_actions() {
-        let temp = TempDir::new().unwrap();
-        let system_store = temp.path().join("system-store");
-        let memory_dir = system_store.join("memory");
-        crate::prompts::ensure_memory_store_layout_at(&memory_dir).unwrap();
-        let store_bindings = crate::AgentRuntimeStoreBindings {
-            rollouts: system_store.join("rollouts"),
-            checkpoints: system_store.join("checkpoints"),
-            cache: system_store.join("cache"),
-            tmp: system_store.join("tmp"),
-            metadata: system_store.join("metadata"),
-        };
-        for path in [
-            &store_bindings.rollouts,
-            &store_bindings.checkpoints,
-            &store_bindings.cache,
-            &store_bindings.tmp,
-            &store_bindings.metadata,
-        ] {
-            std::fs::create_dir_all(path).unwrap();
-        }
-
-        let mut core_config = crate::Config::for_openai_chat_completions_compatible(
-            "sk-test",
-            None,
-            Some("test-model"),
-        );
-        core_config.memory.enabled = true;
-        core_config.memory.store_dir = Some(memory_dir.clone());
-        core_config.streaming_mode = crate::config::StreamingMode::Off;
-
-        let mut agent_config = crate::AgentConfig::from(core_config);
-        agent_config.runtime_config.streaming_mode = crate::config::StreamingMode::Off;
-
-        let config = AgentProcessConfig {
-            agent_config,
-            launch_context: crate::ProcessLaunchContext::root().with_descriptor(
-                crate::MEMORY_STORE_DESCRIPTOR,
-                crate::ProcessDescriptor::new("/memory").unwrap(),
-            ),
-            store_bindings: Some(store_bindings),
-            memory_store_backing: Some(memory_dir.clone()),
-            ..AgentProcessConfig::default()
-        };
-        let call_count = Arc::new(Mutex::new(0));
-        let agentfs = Arc::new(alan_agentfs::AgentFs::new());
-        let llmfs = Arc::new(alan_llmfs::LlmFs::new());
-        llmfs.register_connection(
-            "default",
-            Box::new(ShutdownDrainMemoryPromotionProvider {
-                call_count: Arc::clone(&call_count),
-                deferred_delay: Duration::from_millis(100),
-            }),
-        );
-        let mut namespace = alan_kernel::Namespace::new();
-        namespace.mount(
-            "/agent/1",
-            alan_ap::InProcessTransport::new(agentfs),
-            alan_kernel::Access::ReadWrite,
-        );
-        namespace.mount(
-            "/mnt/llm",
-            alan_ap::InProcessTransport::new(llmfs),
-            alan_kernel::Access::ReadWrite,
-        );
-        let root = alan_ap::InProcessTransport::new(Arc::new(alan_kernel::MountFs::new(namespace)));
-        let shell = alan_shell::Shell::new(root.clone());
-        let mut output = shell.tail("/agent/1/io/output").await.unwrap();
-        let environment =
-            crate::runtime::NamespaceRuntimeEnvironment::new(root, "/agent/1", "default");
-        let generation_capabilities =
-            crate::provider_capabilities_for_config(&config.agent_config.core_config);
-        let mut controller = spawn_with_namespace_environment(
-            config,
-            environment,
-            crate::skills::SkillHostCapabilities::default(),
-            generation_capabilities,
-        )
-        .unwrap();
-        controller.wait_until_ready().await.unwrap();
-
-        let submission = Submission::new(Op::Turn {
-            parts: vec![alan_agent_protocol::ContentPart::text("My name is Morris.")],
-            context: None,
-        });
-        controller
-            .handle
-            .submission_tx
-            .send(submission.clone())
-            .await
-            .unwrap();
-
-        let output = tokio::time::timeout(Duration::from_secs(15), output.read(4096))
-            .await
-            .expect("turn output did not arrive")
-            .unwrap();
-        assert_eq!(String::from_utf8(output).unwrap(), "Noted.");
-
-        controller.shutdown().await.unwrap();
-
-        let user_memory =
-            tokio::fs::read_to_string(memory_dir.join(crate::prompts::MEMORY_USER_FILENAME))
-                .await
-                .unwrap();
-        assert!(
-            user_memory.contains("Name: Morris"),
-            "provider_calls={}, user_memory={user_memory:?}",
-            *call_count.lock().unwrap()
-        );
-    }
-
-    #[tokio::test]
-    async fn test_spawn_with_namespace_environment_reaches_ready_without_store_bindings() {
-        let core_config = crate::Config::default();
-        let generation_capabilities = crate::provider_capabilities_for_config(&core_config);
-        let config = AgentProcessConfig {
-            agent_config: crate::AgentConfig::from(core_config),
-            ..AgentProcessConfig::default()
-        };
-        let mut ns = alan_kernel::Namespace::new();
-        ns.mount(
-            "/agent/1",
-            alan_ap::InProcessTransport::new(Arc::new(alan_agentfs::AgentFs::new())),
-            alan_kernel::Access::ReadWrite,
-        );
-        let root = alan_ap::InProcessTransport::new(Arc::new(alan_kernel::MountFs::new(ns)));
-        let namespace_environment =
-            crate::runtime::NamespaceRuntimeEnvironment::new(root, "/agent/1", "default");
-
-        let mut controller = spawn_with_namespace_environment(
-            config,
-            namespace_environment,
-            crate::skills::SkillHostCapabilities::default(),
-            generation_capabilities,
-        )
-        .unwrap();
-        let ready = controller.wait_until_ready().await.unwrap();
-
-        assert_eq!(ready.process_path, "/proc/1");
-        assert_eq!(ready.agent_path, "/agent/1");
-        assert!(ready.rollout_id.is_none());
-        assert!(!ready.durability.durable);
-        controller.shutdown().await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_namespace_io_input_frame_drives_runtime_turn_without_api_submission() {
-        let agentfs = Arc::new(alan_agentfs::AgentFs::new());
-        let llmfs = Arc::new(alan_llmfs::LlmFs::new());
-        let mock = MockLlmProvider::new().with_responses(vec![
-            GenerationResponse {
-                content: "first namespace response".to_string(),
-                thinking: None,
-                thinking_signature: None,
-                redacted_thinking: Vec::new(),
-                tool_calls: Vec::new(),
-                usage: None,
-                finish_reason: Some("stop".to_string()),
-                provider_response_id: None,
-                provider_response_status: None,
-                warnings: Vec::new(),
-            },
-            GenerationResponse {
-                content: "second namespace response".to_string(),
-                thinking: None,
-                thinking_signature: None,
-                redacted_thinking: Vec::new(),
-                tool_calls: Vec::new(),
-                usage: None,
-                finish_reason: Some("stop".to_string()),
-                provider_response_id: None,
-                provider_response_status: None,
-                warnings: Vec::new(),
-            },
-        ]);
-        let mock_probe = mock.clone();
-        llmfs.register_connection("default", Box::new(mock));
-
-        let procfs = Arc::new(alan_kernel::ProcFs::new());
-        let agent_root = Arc::new(alan_agentfs::AgentRootFs::new(procfs.clone()));
-        let mut ns = alan_kernel::Namespace::new();
-        ns.mount(
-            "/proc",
-            alan_ap::InProcessTransport::new(procfs),
-            alan_kernel::Access::ReadWrite,
-        );
-        ns.mount(
-            "/agent",
-            alan_ap::InProcessTransport::new(agent_root.clone()),
-            alan_kernel::Access::ReadWrite,
-        );
-        ns.mount(
-            "/mnt/llm",
-            alan_ap::InProcessTransport::new(llmfs),
-            alan_kernel::Access::ReadWrite,
-        );
-        for path in ["/bin", "/lib", "/man", "/mnt"] {
-            ns.mount(
-                path,
-                alan_ap::InProcessTransport::new(Arc::new(alan_ap::reference::MemFs::empty())),
-                alan_kernel::Access::ReadOnly,
-            );
-        }
-        for name in ["read_file", "write_file", "search_files", "run_command"] {
-            let manifest = crate::runtime::ToolPackageManifest::from_tool(
-                &PackageTestTool {
-                    name,
-                    description: "Host-mounted test Tool",
-                },
-                30,
-            )
-            .unwrap();
-            ns.mount(
-                &format!("/bin/{name}"),
-                alan_ap::InProcessTransport::new(Arc::new(alan_ap::reference::MemFs::empty())),
-                alan_kernel::Access::ReadOnly,
-            );
-            ns.mount(
-                &format!("/lib/exec/{name}"),
-                alan_ap::InProcessTransport::new(Arc::new(
-                    alan_ap::reference::MemFs::with_read_only_file(
-                        "manifest",
-                        serde_json::to_vec(&manifest).unwrap(),
-                    ),
-                )),
-                alan_kernel::Access::ReadOnly,
-            );
-        }
-        let live_namespace = alan_kernel::LiveNamespace::new(ns);
-        let root = alan_ap::InProcessTransport::new(Arc::new(
-            alan_kernel::MountFs::from_live_namespace(live_namespace.clone()),
-        ));
-        let bootstrap_shell = alan_shell::Shell::new(root.clone());
-        let pid = bootstrap_shell
-            .spawn(r#"{"executable":"/bin/agent","args":[]}"#)
-            .await
-            .unwrap();
-        assert_eq!(pid, "1");
-        agent_root.bind_process(pid.clone(), agentfs).await;
-        agent_root.set_root_process(pid).await;
-        let (client_stream, server_stream) = tokio::io::duplex(64 * 1024);
-        let (client_read, client_write) = tokio::io::split(client_stream);
-        let (server_read, server_write) = tokio::io::split(server_stream);
-        let attachment_mount = Arc::new(alan_kernel::MountFs::from_live_namespace(live_namespace));
-        let server_task = tokio::spawn(alan_ap::export_file_server(
-            attachment_mount,
-            tokio::io::BufReader::new(server_read),
-            server_write,
-        ));
-        let imported = Arc::new(alan_ap::ImportedFileServer::new(
-            tokio::io::BufReader::new(client_read),
-            client_write,
-        ));
-        let shell = alan_shell::Shell::new(alan_ap::InProcessTransport::new(imported));
-
-        let mut core_config = crate::Config::default();
-        core_config.memory.enabled = false;
-        let generation_capabilities = crate::provider_capabilities_for_config(&core_config);
-        let store = TempDir::new().unwrap();
-        let store_bindings = crate::AgentRuntimeStoreBindings {
-            rollouts: store.path().join("rollouts"),
-            checkpoints: store.path().join("checkpoints"),
-            cache: store.path().join("cache"),
-            tmp: store.path().join("tmp"),
-            metadata: store.path().join("metadata"),
-        };
-        for path in [
-            &store_bindings.rollouts,
-            &store_bindings.checkpoints,
-            &store_bindings.cache,
-            &store_bindings.tmp,
-            &store_bindings.metadata,
-        ] {
-            std::fs::create_dir_all(path).unwrap();
-        }
-        let config = AgentProcessConfig {
-            agent_config: crate::AgentConfig::from(core_config),
-            store_bindings: Some(store_bindings),
-            ..AgentProcessConfig::default()
-        };
-        let namespace_environment =
-            crate::runtime::NamespaceRuntimeEnvironment::new(root, "/agent/1", "default");
-        let mut controller = spawn_with_namespace_environment(
-            config,
-            namespace_environment,
-            crate::skills::SkillHostCapabilities::default(),
-            generation_capabilities,
-        )
-        .unwrap();
-        controller.wait_until_ready().await.unwrap();
-
-        let mut ui_events = shell.tail("/agent/1/machine/ui/events").await.unwrap();
-        shell
-            .write("/agent/1/io/input", b"hello through files")
-            .await
-            .unwrap();
-
-        wait_for_ui_turn_completion(&mut ui_events, Duration::from_secs(5)).await;
-
-        let output = String::from_utf8(shell.cat("/agent/1/io/output").await.unwrap()).unwrap();
-        assert_eq!(output, "first namespace response");
-
-        shell
-            .write("/agent/1/io/input", b"second input through files")
-            .await
-            .unwrap();
-        wait_for_ui_turn_completion(&mut ui_events, Duration::from_secs(5)).await;
-        let output = String::from_utf8(shell.cat("/agent/1/io/output").await.unwrap()).unwrap();
-        assert_eq!(output, "first namespace responsesecond namespace response");
-        assert_eq!(mock_probe.recorded_requests().len(), 2);
-
-        controller.shutdown().await.unwrap();
-        drop(shell);
-        server_task.abort();
-    }
-
-    #[tokio::test]
-    async fn test_outer_idle_reads_answered_namespace_request_response() {
-        let agentfs = Arc::new(alan_agentfs::AgentFs::new());
-        let mut ns = alan_kernel::Namespace::new();
-        ns.mount(
-            "/agent/1",
-            alan_ap::InProcessTransport::new(agentfs),
-            alan_kernel::Access::ReadWrite,
-        );
-        let root = alan_ap::InProcessTransport::new(Arc::new(alan_kernel::MountFs::new(ns)));
-        let shell = alan_shell::Shell::new(root.clone());
-        let namespace_environment =
-            crate::runtime::NamespaceRuntimeEnvironment::new(root, "/agent/1", "default");
-
-        let request_id = namespace_environment
-            .write_request(crate::runtime::agent_loop::NamespaceRequestRecord::new(
-                "structured_input",
-                "Provide the missing value",
-            ))
-            .await
-            .unwrap();
-        let mut turn_state = TurnState::default();
-        turn_state.set_structured_input(crate::approval::PendingStructuredInputRequest {
-            request_id: request_id.clone(),
-            title: "Missing value".to_string(),
-            prompt: "Provide the missing value".to_string(),
-            questions: Vec::new(),
-        });
-        let state = RuntimeLoopState {
-            machine: crate::AgentMachine::new(),
-            current_submission_id: None,
-            environment: namespace_environment,
-            core_config: crate::Config::default(),
-            runtime_config: RuntimeConfig::default(),
-            definition_persona_dirs: Vec::new(),
-            prompt_cache: crate::runtime::prompt_cache::PromptAssemblyCache::new(Vec::new()),
-            turn_state,
-        };
-
-        assert!(
-            read_pending_namespace_resume_submission(&state)
-                .await
-                .is_none(),
-            "unanswered request should not create a resume submission"
-        );
-
-        shell
-            .write(
-                &format!("/agent/1/requests/{request_id}/response"),
-                br#"{"answers":[{"question_id":"q1","value":"from file"}]}"#,
-            )
-            .await
-            .unwrap();
-
-        let submission = read_pending_namespace_resume_submission(&state)
-            .await
-            .expect("answered namespace request should be observed")
-            .unwrap();
-        match submission.op {
-            Op::Resume {
-                request_id: resumed_id,
-                content,
-            } => {
-                assert_eq!(resumed_id, request_id);
-                assert_eq!(
-                    content,
-                    vec![ContentPart::structured(serde_json::json!({
-                        "answers": [{"question_id": "q1", "value": "from file"}]
-                    }))]
-                );
-            }
-            other => panic!("expected Op::Resume from namespace response, got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn test_namespace_machine_ctl_drives_runtime_submission_without_api_submission() {
-        let agentfs = Arc::new(alan_agentfs::AgentFs::new());
-        let llmfs = Arc::new(alan_llmfs::LlmFs::new());
-        llmfs.register_connection(
-            "default",
-            Box::new(MockLlmProvider::new().with_response(GenerationResponse {
-                content: "hello from namespace runtime".to_string(),
-                thinking: None,
-                thinking_signature: None,
-                redacted_thinking: Vec::new(),
-                tool_calls: Vec::new(),
-                usage: None,
-                finish_reason: Some("stop".to_string()),
-                provider_response_id: None,
-                provider_response_status: None,
-                warnings: Vec::new(),
-            })),
-        );
-
-        let mut ns = alan_kernel::Namespace::new();
-        ns.mount(
-            "/agent/1",
-            alan_ap::InProcessTransport::new(agentfs),
-            alan_kernel::Access::ReadWrite,
-        );
-        ns.mount(
-            "/mnt/llm",
-            alan_ap::InProcessTransport::new(llmfs),
-            alan_kernel::Access::ReadWrite,
-        );
-        let root = alan_ap::InProcessTransport::new(Arc::new(alan_kernel::MountFs::new(ns)));
-        let shell = alan_shell::Shell::new(root.clone());
-
-        let core_config = crate::Config::default();
-        let generation_capabilities = crate::provider_capabilities_for_config(&core_config);
-        let config = AgentProcessConfig {
-            agent_config: crate::AgentConfig::from(core_config),
-            ..AgentProcessConfig::default()
-        };
-        let namespace_environment =
-            crate::runtime::NamespaceRuntimeEnvironment::new(root, "/agent/1", "default");
-        let mut controller = spawn_with_namespace_environment(
-            config,
-            namespace_environment,
-            crate::skills::SkillHostCapabilities::default(),
-            generation_capabilities,
-        )
-        .unwrap();
-        controller.wait_until_ready().await.unwrap();
-
-        let mut ui_events = shell.tail("/agent/1/machine/ui/events").await.unwrap();
-        shell
-            .write("/agent/1/io/input", b"hello through files")
-            .await
-            .unwrap();
-
-        wait_for_ui_turn_completion(&mut ui_events, Duration::from_secs(5)).await;
-
-        shell
-            .write("/agent/1/machine/ctl", b"rollback")
-            .await
-            .unwrap();
-
-        let rollback_notice = tokio::time::timeout(Duration::from_secs(5), async {
-            let mut pending = String::new();
-            'events: loop {
-                pending.push_str(&String::from_utf8(ui_events.read(4096).await.unwrap()).unwrap());
-                while let Some(newline) = pending.find('\n') {
-                    let line = pending[..newline].to_string();
-                    pending.drain(..=newline);
-                    let event: alan_agent_protocol::UiEvent = serde_json::from_str(&line).unwrap();
-                    if let alan_agent_protocol::UiEvent::Notice { snapshot } = event
-                        && snapshot.kind == alan_agent_protocol::UiNoticeKind::Rollback
-                    {
-                        break 'events snapshot.message;
-                    }
-                }
-            }
-        })
-        .await
-        .expect("namespace machine/ctl should drive rollback submission");
-        assert_eq!(rollback_notice, "rolled back 1 turns");
-
-        controller.shutdown().await.unwrap();
-    }
-
-    #[test]
-    fn test_agent_runtime_config_default() {
-        let config = AgentProcessConfig::default();
-        assert_eq!(config.launch_context.cwd, "/");
-        assert!(config.launch_context.host_mounts.is_empty());
-        assert!(config.launch_context.descriptors.is_empty());
-        assert!(config.store_bindings.is_none());
-        assert!(config.memory_store_backing.is_none());
-    }
-
-    #[test]
-    fn runtime_tool_binding_uses_host_mount_when_process_cwd_is_virtual() {
-        let temp = TempDir::new().unwrap();
-        let source = temp.path().join("source");
-        std::fs::create_dir_all(&source).unwrap();
-        let mut launch_context = crate::ProcessLaunchContext::root();
-        launch_context.namespace.mount(
-            "/mnt/source",
-            InProcessTransport::new(Arc::new(alan_ap::reference::MemFs::new())),
-            alan_kernel::Access::ReadWrite,
-        );
-        launch_context = launch_context.with_host_mount(
-            crate::HostMountGrant::new("/mnt/source", &source, alan_kernel::Access::ReadWrite)
-                .unwrap(),
-        );
-        let store_root = temp.path().join("system-store");
-        let config = AgentProcessConfig {
-            launch_context,
-            store_bindings: Some(crate::AgentRuntimeStoreBindings {
-                rollouts: store_root.join("rollouts"),
-                checkpoints: store_root.join("checkpoints"),
-                cache: store_root.join("cache"),
-                tmp: store_root.join("tmp"),
-                metadata: store_root.join("metadata"),
-            }),
-            ..AgentProcessConfig::default()
-        };
-        let mut tools = crate::tools::ToolRegistry::new();
-
-        configure_runtime_tool_execution_binding(&config, &mut tools).unwrap();
-
-        let binding = tools
-            .default_execution_binding()
-            .expect("an explicit Host Mount must create a runtime Tool binding");
-        assert_eq!(binding.cwd, dunce::canonicalize(&source).unwrap());
-        assert_eq!(binding.namespace_cwd, PathBuf::from("/mnt/source"));
-        assert_eq!(config.launch_context.cwd, "/");
-        assert_eq!(binding.host_mounts, config.launch_context.host_mounts);
-    }
-
-    #[test]
-    fn package_projection_alone_does_not_require_runtime_tool_binding() {
-        let mut launch_context = crate::ProcessLaunchContext::root();
-        launch_context.add_package_reference(
-            crate::ProcessPackageReference::new(
-                "example",
-                "a".repeat(64),
-                crate::ProcessPackageKind::Installed,
-                "/lib/pkg/example",
-                Vec::new(),
-                alan_ap::InProcessTransport::new(std::sync::Arc::new(
-                    alan_ap::reference::MemFs::new(),
-                )),
-            )
-            .unwrap(),
-        );
-        let config = AgentProcessConfig {
-            launch_context,
-            store_bindings: None,
-            ..AgentProcessConfig::default()
-        };
-        let mut tools = crate::tools::ToolRegistry::new();
-
-        configure_runtime_tool_execution_binding(&config, &mut tools).unwrap();
-
-        assert!(tools.default_execution_binding().is_none());
-    }
-
-    #[test]
-    fn test_runtime_host_capabilities_enable_delegated_support_for_top_level_runtime() {
-        let config = AgentProcessConfig::default();
-        let tools = crate::tools::ToolRegistry::new();
-
-        let capabilities = runtime_host_capabilities(&config, &tools);
-
-        assert!(capabilities.supports_delegated_skill_invocation());
-        assert!(capabilities.tools.contains("invoke_delegated_skill"));
-    }
-
-    #[test]
-    fn test_runtime_host_capabilities_include_host_path_executables() {
-        let temp = tempfile::TempDir::new().unwrap();
-        let executable_path = {
-            #[cfg(windows)]
-            {
-                temp.path().join("demo.cmd")
-            }
-
-            #[cfg(not(windows))]
-            {
-                temp.path().join("demo")
-            }
-        };
-        std::fs::write(&executable_path, "echo demo\n").unwrap();
-
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-
-            let mut permissions = std::fs::metadata(&executable_path).unwrap().permissions();
-            permissions.set_mode(0o755);
-            std::fs::set_permissions(&executable_path, permissions).unwrap();
-        }
-
-        let capabilities = runtime_host_capabilities_with_path_dirs(
-            &AgentProcessConfig::default(),
-            &crate::tools::ToolRegistry::new(),
-            [temp.path()],
-        );
-
-        assert!(capabilities.supports_required_tool("demo"));
-    }
-
-    #[test]
-    fn test_agent_runtime_config_from_core_config() {
-        let core_config = crate::config::Config::default();
-        let runtime_config = AgentProcessConfig::from(core_config);
-
-        assert_eq!(runtime_config.launch_context.cwd, "/");
-        assert!(runtime_config.launch_context.host_mounts.is_empty());
-        assert!(runtime_config.store_bindings.is_none());
-    }
-
-    #[test]
-    fn test_agent_runtime_config_clone() {
-        let config = AgentProcessConfig::default();
-        let cloned = config.clone();
-        assert_eq!(config.launch_context.cwd, cloned.launch_context.cwd);
-        assert_eq!(
-            config.launch_context.host_mounts,
-            cloned.launch_context.host_mounts
-        );
-    }
-
-    #[test]
-    fn test_agent_runtime_config_debug() {
-        let config = AgentProcessConfig::default();
-        let debug_str = format!("{:?}", config);
-        assert!(debug_str.contains("AgentProcessConfig"));
-        assert!(debug_str.contains("launch_context"));
-        assert!(!debug_str.contains("workspace_id"));
-    }
-
-    #[test]
-    fn test_agent_runtime_handle_clone() {
-        let (sub_tx, _sub_rx) = mpsc::channel(10);
-
-        let handle = RuntimeHandle {
-            submission_tx: sub_tx,
-            shutdown_tx: None,
-        };
-
-        let cloned = handle.clone();
-        // Both handles should share the same channels
-        drop(cloned);
-        drop(handle);
-    }
-
-    #[test]
-    fn test_agent_runtime_handle_fields() {
-        let (sub_tx, _sub_rx) = mpsc::channel::<Submission>(10);
-
-        let handle = RuntimeHandle {
-            submission_tx: sub_tx,
-            shutdown_tx: None,
-        };
-
-        // Verify handle can be created
-        assert!(!handle.submission_tx.is_closed());
-    }
-
-    #[tokio::test]
-    async fn test_agent_runtime_handle_shutdown_without_channel() {
-        let (sub_tx, _sub_rx) = mpsc::channel::<Submission>(10);
-        let handle = RuntimeHandle {
-            submission_tx: sub_tx,
-            shutdown_tx: None,
-        };
-
-        let result = handle.shutdown().await;
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_agent_config_with_definition_overlays_updates_unmodified_runtime_fields() {
-        let temp = TempDir::new().unwrap();
-        let overlay_path = temp.path().join("agent.toml");
-        write_agent_overlay(
-            &overlay_path,
-            r#"
-	tool_repeat_limit = 9
-	prompt_snapshot_enabled = true
-	"#,
-        );
-
-        let base = AgentConfig::from(crate::Config::default());
-        let merged = base.with_definition_overlays(&[overlay_path]).unwrap();
-
-        assert_eq!(merged.core_config.tool_repeat_limit, 9);
-        assert!(merged.core_config.prompt_snapshot_enabled);
-        assert_eq!(merged.runtime_config.tool_repeat_limit, 9);
-        assert!(merged.runtime_config.prompt_snapshot_enabled);
-    }
-
-    #[test]
-    fn test_agent_config_with_definition_overlays_updates_unmodified_reasoning_effort() {
-        let temp = TempDir::new().unwrap();
-        let overlay_path = temp.path().join("agent.toml");
-        write_agent_overlay(
-            &overlay_path,
-            r#"
-model_reasoning_effort = "high"
-"#,
-        );
-
-        let base = AgentConfig::from(crate::Config::default());
-        let merged = base.with_definition_overlays(&[overlay_path]).unwrap();
-
-        assert_eq!(
-            merged.core_config.model_reasoning_effort,
-            Some(alan_agent_protocol::ReasoningEffort::High)
-        );
-        assert_eq!(
-            merged
-                .runtime_config
-                .request_control_intent
-                .reasoning_effort,
-            Some(alan_agent_protocol::ReasoningEffort::High)
-        );
-    }
-
-    #[test]
-    fn test_agent_config_with_definition_overlays_preserves_runtime_overrides() {
-        let temp = TempDir::new().unwrap();
-        let overlay_path = temp.path().join("agent.toml");
-        write_agent_overlay(
-            &overlay_path,
-            r#"
-	tool_repeat_limit = 9
-	streaming_mode = "off"
-	model_reasoning_effort = "high"
-	"#,
-        );
-
-        let mut base = AgentConfig::from(crate::Config::default());
-        base.runtime_config.tool_repeat_limit = 42;
-        base.set_model_override("gpt-5-mini");
-        base.set_streaming_mode_override(crate::config::StreamingMode::On);
-        base.set_model_reasoning_effort_override(Some(alan_agent_protocol::ReasoningEffort::Low));
-
-        let merged = base.with_definition_overlays(&[overlay_path]).unwrap();
-
-        assert_eq!(merged.core_config.openai_responses_model, "gpt-5-mini");
-        assert_eq!(merged.core_config.tool_repeat_limit, 9);
-        assert_eq!(
-            merged.core_config.streaming_mode,
-            crate::config::StreamingMode::On
-        );
-        assert_eq!(
-            merged.core_config.model_reasoning_effort,
-            Some(alan_agent_protocol::ReasoningEffort::Low)
-        );
-        assert_eq!(
-            merged.core_config.effective_context_window_tokens(),
-            crate::Config::for_openai_responses("sk-test", None, Some("gpt-5-mini"))
-                .effective_context_window_tokens()
-        );
-        assert_eq!(merged.runtime_config.tool_repeat_limit, 42);
-        assert_eq!(
-            merged.runtime_config.context_window_tokens,
-            crate::Config::for_openai_responses("sk-test", None, Some("gpt-5-mini"))
-                .effective_context_window_tokens()
-        );
-        assert_eq!(
-            merged.runtime_config.streaming_mode,
-            crate::config::StreamingMode::On
-        );
-        assert_eq!(
-            merged
-                .runtime_config
-                .request_control_intent
-                .reasoning_effort,
-            Some(alan_agent_protocol::ReasoningEffort::Low)
-        );
-    }
-
-    #[test]
-    fn test_set_model_override_refreshes_runtime_context_window_budget() {
-        let mut config = AgentConfig::from(crate::Config::for_openai_responses(
-            "sk-test",
-            None,
-            Some("gpt-5.4"),
-        ));
-        assert_eq!(config.runtime_config.context_window_tokens, 1_050_000);
-
-        config.set_model_override("gpt-5-mini");
-
-        assert_eq!(config.core_config.effective_model(), "gpt-5-mini");
-        assert_eq!(config.runtime_config.context_window_tokens, 400_000);
-    }
-
-    #[test]
-    fn test_agent_config_with_definition_overlays_preserves_marked_same_value_runtime_overrides() {
-        let temp = TempDir::new().unwrap();
-        let overlay_path = temp.path().join("agent.toml");
-        write_agent_overlay(
-            &overlay_path,
-            r#"
-streaming_mode = "off"
-partial_stream_recovery_mode = "off"
-[durability]
-required = true
-"#,
-        );
-
-        let mut base = AgentConfig::from(crate::Config::default());
-        base.set_streaming_mode_override(crate::config::StreamingMode::Auto);
-        base.set_partial_stream_recovery_mode_override(
-            crate::config::PartialStreamRecoveryMode::ContinueOnce,
-        );
-        base.set_durability_required_override(false);
-
-        let merged = base.with_definition_overlays(&[overlay_path]).unwrap();
-
-        assert_eq!(
-            merged.core_config.streaming_mode,
-            crate::config::StreamingMode::Auto
-        );
-        assert_eq!(
-            merged.runtime_config.streaming_mode,
-            crate::config::StreamingMode::Auto
-        );
-        assert_eq!(
-            merged.core_config.partial_stream_recovery_mode,
-            crate::config::PartialStreamRecoveryMode::ContinueOnce
-        );
-        assert_eq!(
-            merged.runtime_config.partial_stream_recovery_mode,
-            crate::config::PartialStreamRecoveryMode::ContinueOnce
-        );
-        assert!(!merged.core_config.durability.required);
-        assert!(!merged.runtime_config.durability_required);
-    }
-
-    #[tokio::test]
-    async fn test_agent_runtime_handle_shutdown_with_channel() {
-        let (sub_tx, _sub_rx) = mpsc::channel::<Submission>(10);
-        let (shutdown_tx, mut shutdown_rx) = mpsc::channel::<()>(1);
-
-        let handle = RuntimeHandle {
-            submission_tx: sub_tx,
-            shutdown_tx: Some(shutdown_tx),
-        };
-
-        // Shutdown should send signal
-        let result = handle.shutdown().await;
-        assert!(result.is_ok());
-
-        // Verify shutdown signal was sent
-        let signal = shutdown_rx.recv().await;
-        assert!(signal.is_some());
-    }
-
-    #[test]
-    fn test_agent_runtime_config_recovery_rollout_path() {
-        let temp = TempDir::new().unwrap();
-        let rollout_path = temp.path().join("rollout.jsonl");
-
-        let config = AgentProcessConfig {
-            recovery_rollout_path: Some(rollout_path.clone()),
-            ..Default::default()
-        };
-
-        assert_eq!(config.recovery_rollout_path, Some(rollout_path));
-    }
-
-    #[tokio::test]
-    async fn test_initialize_agent_machine_from_rollout_preserves_current_process_cwd() {
-        let temp = TempDir::new().unwrap();
-        let process_cwd = std::path::Path::new("/mnt/source/src");
-        let recovered_rollouts = temp.path().join("recovered-rollouts");
-        let mut source =
-            AgentMachine::new_with_recorder_in_dir("/proc/41", "gemini-2.0-flash", temp.path())
-                .await
-                .unwrap();
-        source.add_user_message("Hello");
-        source.flush().await;
-        let rollout_path = source.rollout_path().unwrap().clone();
-        drop(source);
-
-        let startup = initialize_agent_machine(
-            AgentMachineLaunchContext {
-                process_path: "/proc/42",
-                agent_path: "/agent/42",
-                model: "gemini-2.0-flash",
-            },
-            Some(&rollout_path),
-            Some(&recovered_rollouts),
-            true,
-            Some(process_cwd),
-            crate::ResolvedRequestControls {
-                reasoning: alan_agent_protocol::ReasoningControls {
-                    effort: Some(alan_agent_protocol::ReasoningEffort::Medium),
-                },
-                source: crate::RequestControlSource::AgentMachineOverride,
-                diagnostics: Vec::new(),
-            },
-        )
-        .await
-        .unwrap();
-
-        assert_eq!(startup.metadata.process_path, "/proc/42");
-        assert_eq!(startup.metadata.agent_path, "/agent/42");
-        assert!(startup.metadata.rollout_id.is_some());
-
-        let persisted_path = startup
-            .metadata
-            .rollout_path
-            .clone()
-            .expect("recovered machine should create a new rollout recorder");
-        let persisted_items = crate::rollout::RolloutRecorder::load_history(&persisted_path)
-            .await
-            .unwrap();
-        let persisted_meta = persisted_items.into_iter().find_map(|item| match item {
-            crate::rollout::RolloutItem::AgentMachineMeta(meta) => Some(meta),
-            _ => None,
-        });
-
-        assert_eq!(
-            persisted_meta.as_ref().map(|meta| meta.cwd.as_str()),
-            Some("/mnt/source/src")
-        );
-        assert_eq!(
-            persisted_meta
-                .as_ref()
-                .map(|meta| meta.process_path.as_str()),
-            Some("/proc/42")
-        );
-        assert_eq!(
-            persisted_meta
-                .as_ref()
-                .and_then(|meta| meta.reasoning_effort),
-            Some(alan_agent_protocol::ReasoningEffort::Medium)
-        );
-
-        drop(startup);
-        let _ = tokio::fs::remove_file(persisted_path).await;
-    }
-
-    #[test]
-    fn test_should_drive_turn_submission() {
-        // steer/follow_up should be driven as turn
-        assert!(should_drive_turn_submission(&Op::Input {
-            parts: vec![alan_agent_protocol::ContentPart::text("test")],
-            mode: alan_agent_protocol::InputMode::Steer,
-        }));
-        assert!(should_drive_turn_submission(&Op::Input {
-            parts: vec![alan_agent_protocol::ContentPart::text("test")],
-            mode: alan_agent_protocol::InputMode::FollowUp,
-        }));
-        // next_turn should be queue-only, not immediate execution.
-        assert!(!should_drive_turn_submission(&Op::Input {
-            parts: vec![alan_agent_protocol::ContentPart::text("test")],
-            mode: alan_agent_protocol::InputMode::NextTurn,
-        }));
-
-        // Turn should be driven as turn
-        assert!(should_drive_turn_submission(&Op::Turn {
-            parts: vec![alan_agent_protocol::ContentPart::text("test")],
-            context: None,
-        }));
-
-        // Other ops should not be driven as turn
-        assert!(!should_drive_turn_submission(&Op::CompactWithOptions {
-            focus: None,
-        }));
-        assert!(!should_drive_turn_submission(&Op::CompactWithOptions {
-            focus: Some("preserve todos".to_string()),
-        }));
-        assert!(!should_drive_turn_submission(&Op::Rollback { turns: 1 }));
-        assert!(!should_drive_turn_submission(&Op::Interrupt));
-        assert!(!should_drive_turn_submission(&Op::Resume {
-            request_id: "req-123".to_string(),
-            content: vec![alan_agent_protocol::ContentPart::structured(
-                serde_json::json!({})
-            )],
-        }));
-    }
-}
+#[path = "engine_tests.rs"]
+mod tests;
