+++ scripts/rust-source-size-baseline.txt
@@ -4,7 +4,7 @@ crates/agent-engine/skills/swebench/tooling/src/swebench_lite.rs 1389
 crates/agent-engine/src/agent_machine.rs 3661
 crates/agent-engine/src/config.rs 2556
 crates/agent-engine/src/rollout.rs 2266
-crates/agent-engine/src/runtime/engine.rs 2816
+crates/agent-engine/src/runtime/engine.rs 1378
 crates/agent-engine/src/skills/injector.rs 2490
 crates/agent-engine/src/skills/types.rs 3013
 crates/agent-engine/src/tools/reified_namespace.rs 2756
